ibtop: Add main() function
ibtop is a curses application for monitoring of the communication
hotspots on the HCA side. ibtop does not look at switches.
This first commit is a rudimentary version of ibtop. It displays
the rx and tx bandwidth in units of Mbps and the percentage of
the peak (ibtop is developed on an FDR cluster so the theoretical
peak is currently hardcoded).