ibtop: Sort nodes by bandwidth
The current code version uses the order of the nodes as obtained
from libibnetdiscover. In order to increase the visual
representation of the fabric load this commit modifies the code
such that nodes are sorted according to the maximum tx/rx bandwidth.
In order to avoid unnecessary jitter the node order is kept fixed
over a period of window updates (10 in the current setup).