Add ibutil006 and cleanup code
This commit adds ibutil006 which maps GUIDs to LIDs by means of
PathRecord MADs. The same functionality is available via the
Python interface.

At the same time a number of bugs and quirks of the existing code
are removed. In particular, we do not reuse transaction ids as
before in order to make sure that sending sequences of MADs works
properly.